chore #31: 함께 달리기 상세 조회 UI 구현

+++ android/core/model/src/main/java/com/sixkids/model/Group.kt
@@ -1,7 +1,7 @@
 package com.sixkids.model
 
 data class Group(
-    val id: Int = 0,
-    val leaderId: Int = 0,
+    val id: Long = 0,
+    val leaderId: Long = 0,
     val studentList : List<MemberSimple> = emptyList(),
 )
